backend/requirement.txt
absl-py==0.11.0
amqp==5.0.6
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
autopep8==1.5.3
billiard==3.6.4.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.1
celery==5.0.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.7.1
demjson==2.2.4
distlib==0.3.1
Django==3.1.5
django-cors-headers==3.4.0
django-mssql-backend==2.8.
django-redis==4.12.1
django-redis-cache==3.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
google==3.0.0
google-auth==1.4.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
isort==5.5.1
jdcal==1.4.1
Jinja2==2.11.3
jsonschema==3.2.0
kombu==5.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
minio==7.0.3
mistune==0.8.4
nest-asyncio==1.5.1
packaging==20.9
pandocfilters==1.4.3
pickleshare==0.7.5
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.6.0
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2020.1
requests==2.21.0
rsa==4.7.2
simplegeneric==0.8.1
six==1.12.0
soupsieve==2.0.1
sqlparse==0.3.1
testpath==0.4.4
toml==0.10.1
tornado==4.5.3
traitlets==5.0.5
typed-ast==1.4.1
urllib3==1.24.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wfastcgi==3.0.0
wrapt==1.12.1
mysqlclient==2.0.3